Fix GH integration for filenames with collapsable whitespace
Since the GitHub integration retrieves filenames directly from the DOM, and HTML collapses sequences of repeated whitespace, it was unable to load files whose names contained such whitespace sequences. Switch to a whitespace-preserving setting for `white-space` so the whitespace will be preserved.

This has the side effect of also showing all whitespace in filenames to users, which is probably also a good thing.