Fix top-right controls being forced offscreen for long filenames
When the name of the current file is too long, it forces the controls in
the top-right of the menubar offscreen (language, settings, facsimile
and annotations panels, etc.) rendering them unclickable.

This change truncates the filename with a continuation indicator (...).
On mouseover or focus, the full filename shown while still allowing the
top-right buttons to be clicked.

Now that the right-hand edge of the file name area is no longer reliably
the end of the filename, it looks visually confusing to have the
file-changed-indicator there. This change moves it to the left-hand side
of the filename.